requirements.txt
streamlit
streamlit-drawable-canvas
PyYAML
pillow
